+++ ChangeLog.md
@@ -1,6 +1,8 @@
 ## ChangeLog - Most recent changes at the top
 
-Version 1.603
+Version 1.603 - Override Function support for require files
+- FEATURE: Provided support for Override require files to override specific functions without changing the default require file.
+To Use: Uncomment the line at the bottom of the default require file, then put things in the custom/override file.  See files for more info.
 - hudclass.lua FEATURE: Changed fuel tank display to color code bars by type and to put a small gap between types. (Thanks Zrips!)
 - hudclass.lua FIX: Fixed issue with buttons not being showable if in keyboard mode and freelook toggle is off.  (Hold alt-shift to see buttons)
 - Removed user variable `WipeDamage` and its code support since a repair unit can roll back a databank to last autosnapshot.